requirements.txt
numpy==1.24.3
flask==3.0.0
openai-whisper==20231117
torch==2.1.2
werkzeug==3.0.1
